requirements.txt
fastapi==0.115.11
pillow==11.1.0
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0